requirements.txt
playwright>=1.50.0
Pillow>=12.0.0
